Move file methods to new classes/module
